one repositoryadapter to rule them all

+++ src/main/java/com/gitgudgang/dogeverse/repository/DogJpaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DogJpaRepository extends CrudRepository<DogEntity, Integer> {
+public interface DogJpaRepository extends CrudRepository<DogEntity, Long> {
 }
